pyproject.toml
[project]
name = "deep-research-studio"
version = "0.1.0"
description = "LangGraph + MCP powered deep research workflow with Streamlit UI"
authors = [{ name = "Your Name" }]
requires-python = ">=3.10"
dependencies = [
    "langchain>=0.3.0",
    "langchain-openai>=0.1.23",
    "langchain-community>=0.3.0",
    "langgraph>=0.2.34",
    "langgraph-sdk>=0.1.34",
    "streamlit>=1.38.0",
    "pydantic>=2.8.0",
    "pydantic-settings>=2.4.0",
    "python-dotenv>=1.0.1",
    "tavily-python>=0.3.3",
    "requests>=2.32.3",
    "mcp>=0.2.3",
    "tenacity>=8.5.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "ruff>=0.5.6"
]

[build-system]
requires = ["setuptools>=69.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
